fix static cast issue for clang

+++ src/runtime/ghex_comm/run.cpp
@@ -215,7 +215,7 @@ public:
         m_global_resolution{global_resolution.x, global_resolution.y},
         m_context{
             std::make_unique<context_t>(MPI_COMM_WORLD, (num_sub_domains > 1))},
-        m_barrier{*m_context, num_sub_domains} {
+        m_barrier{*m_context, static_cast<std::size_t>(num_sub_domains)} {
     MPI_Comm_size(MPI_COMM_WORLD, &m_size);
     MPI_Comm_rank(MPI_COMM_WORLD, &m_rank);
